backend/pyproject.toml
[project]
name = "investment-analyzer-backend"
version = "0.1.0"
description = "Investment Analyzer Backend"
authors = [
    {name = "IbeanI",email = "sig.danielzito@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "alembic (>=1.18.0,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "yfinance (>=0.2.40,<1.0.0)",
    "tenacity (>=9.1.2,<10.0.0)",
    "python-multipart (>=0.0.21,<0.0.22)",
    "matplotlib (>=3.10.8,<4.0.0)",
    "streamlit (>=1.53.0,<2.0.0)",
    "plotly (>=6.5.2,<7.0.0)",
    "slowapi (>=0.1.9,<1.0.0)",
    "python-jose[cryptography] (>=3.3.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "bcrypt (>=4.0.0,<5.0.0)",
    "httpx (>=0.27.0,<1.0.0)",
    "itsdangerous (>=2.2.0,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[dependency-groups]
dev = [
    "jupyterlab (>=4.5.2,<5.0.0)",
    "ipywidgets (>=8.1.8,<9.0.0)",
    "pytest (>=8.0.0,<9.0.0)",
    "pytest-cov (>=4.1.0,<5.0.0)",
]

[tool.pytest.ini_options]
# Test configuration
# Note: ENVIRONMENT=test is set in conftest.py before app imports
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
